pyproject.toml
[project]
name = "trakker-ia-benchmark"
version = "0.1.0"
description = ""
authors = [
    {name = "Pedro Henrique",email = "pedrohesm@gmail.com"}
]
license = {text = "GNU"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.119.0,<0.120.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "prophet (>=1.1.7,<2.0.0)",
    "ruff (>=0.14.1,<0.15.0)",
    "anthropic (>=0.71.0,<0.72.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "perplexityai (>=0.17.0,<0.18.0)",
    "groq (>=0.28.0,<0.29.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.24.0,<0.25.0)",
    "numpy (>=2.4.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
